requirements.txt
fastapi
uvicorn
google-generativeai
google-auth
google-api-python-client
httpx
pydantic
python-multipart
paho-mqtt
python-dotenv
openai
anthropic